implement optimized CSparseFloatMatrixBody's buffers growing, compile and test

+++ NeoML/include/NeoML/TraditionalML/StratifiedCrossValidationSubProblem.h
@@ -54,8 +54,8 @@ private:
 	CArray< CArray<int> > objectsPerPart; // the list of object indices for each of the parts
 	int minPartSize; // the minimum number of objects in each part (the total number of elements / the number of parts)
 	int objectsBeforeTestPart; // the number of objects before the test part
-	CArray<int> pointerB; // the pointers to the vector beginnings
-	CArray<int> pointerE; // the pointers to the vector ends
+	CArray<size_t> pointerB; // the pointers to the vector beginnings
+	CArray<size_t> pointerE; // the pointers to the vector ends
 	CFloatMatrixDesc matrix; // the problem matrix descriptor
 
 	int translateIndex( int index ) const;
